Improve formatting, and description of options in setup step
Still a long way to go...
