+++ src/pages/HospexSetup/index.ts
@@ -1,0 +1,1 @@
+export { HospexSetup } from './HospexSetup'
